fix vochainGenesis flag, update testsuit genesis and remove sleep from vochaintest

+++ service/vochain.go
@@ -45,30 +45,46 @@ func Vochain(vconfig *config.VochainCfg, dev, results bool, ma *metrics.Agent) (
 	log.Infof("vochain listening on: %s", vconfig.P2PListen)
 	log.Infof("vochain exposed IP address: %s", vconfig.PublicAddr)
 
-	if dev {
-		filepath := vconfig.DataDir + "/config/genesis.json"
-		if _, err = os.Stat(filepath); os.IsNotExist(err) {
-			log.Debug("genesis does not exist, using hardcoded genesis")
-		} else {
-			var genesisBytes []byte
-			if genesisBytes, err = ioutil.ReadFile(filepath); err != nil {
-				return
-			}
-			log.Debug("found genesis file, comparing with hardcoded genesis")
-			// compare genesis
-			if string(genesisBytes) != vochain.DevelopmentGenesis1 {
-				log.Warn("genesis found is different from the hardcoded genesis, cleaning and restarting vochain")
-				if err = os.RemoveAll(vconfig.DataDir); err != nil {
+	// Genesis file
+	var genesisBytes []byte
+
+	// If genesis file from flag, prioritize it
+	if len(vconfig.Genesis) > 0 {
+		log.Infof("using custom genesis from %s", vconfig.Genesis)
+		if _, err = os.Stat(vconfig.Genesis); os.IsNotExist(err) {
+			return
+		}
+		if genesisBytes, err = ioutil.ReadFile(vconfig.Genesis); err != nil {
+			return
+		}
+	} else {
+		// If genesis flag not defined, use either dev or release genesis
+		if dev {
+			// If dev mode enabled, auto-update genesis file
+			filepath := vconfig.DataDir + "/config/genesis.json"
+			if _, err = os.Stat(filepath); os.IsNotExist(err) {
+				log.Debug("genesis does not exist, using hardcoded genesis")
+			} else {
+				if genesisBytes, err = ioutil.ReadFile(filepath); err != nil {
 					return
 				}
-			} else {
-				log.Debug("genesis are the same, you have the latest dev genesis")
+				log.Debug("found genesis file, comparing with hardcoded genesis")
+				// compare genesis
+				if string(genesisBytes) != vochain.DevelopmentGenesis1 {
+					log.Warn("genesis found is different from the hardcoded genesis, cleaning and restarting vochain")
+					if err = os.RemoveAll(vconfig.DataDir); err != nil {
+						return
+					}
+				} else {
+					log.Debug("genesis are the same, you have the latest dev genesis")
+				}
 			}
+			genesisBytes = []byte(vochain.DevelopmentGenesis1)
+		} else {
+			genesisBytes = []byte(vochain.ReleaseGenesis1)
 		}
-		vnode = vochain.NewVochain(vconfig, []byte(vochain.DevelopmentGenesis1))
-	} else {
-		vnode = vochain.NewVochain(vconfig, []byte(vochain.ReleaseGenesis1))
 	}
+	vnode = vochain.NewVochain(vconfig, genesisBytes)
 	// Scrutinizer
 	if results {
 		log.Info("creating vochain scrutinizer service")
